Fix accessibility duration typing and locale resolution

+++ frontend/flutter_app/lib/design_system/utils/accessibility.dart
@@ -18,7 +18,7 @@ class AccessibilityUtils {
     BuildContext context, {
     int milliseconds = 180,
   }) {
-    final clamped = milliseconds.clamp(0, 200);
+    final clamped = milliseconds.clamp(0, 200).toInt();
     if (reduceMotion(context)) {
       return Duration.zero;
     }
